+++ crates/sdk-core-c-bridge/src/client.rs
@@ -1543,14 +1543,13 @@ impl From<&ClientDnsLoadBalancingOptions> for temporalio_client::DnsLoadBalancin
 
 impl From<&ClientHttpConnectProxyOptions> for temporalio_client::proxy::HttpConnectProxyOptions {
     fn from(opts: &ClientHttpConnectProxyOptions) -> Self {
-        temporalio_client::proxy::HttpConnectProxyOptions {
-            target_addr: opts.target_host.to_string(),
-            basic_auth: if opts.username.size != 0 && opts.password.size != 0 {
+        temporalio_client::proxy::HttpConnectProxyOptions::new(opts.target_host.to_string())
+            .maybe_basic_auth(if opts.username.size != 0 && opts.password.size != 0 {
                 Some((opts.username.to_string(), opts.password.to_string()))
             } else {
                 None
-            },
-        }
+            })
+            .build()
     }
 }